+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to [ChaiLove](https://github.com/RobLoach/ChaiLove) will be 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.22.0 - 2018-05-06
+### Features
+- `love.filesystem.createDirectory()` to make directories
+- `love.filesystem.write()` to write a file in the save directory
+- `Http()` to make HTTP requests
+
+### Fixes
+- Fix debug build due to undefined "z_error" and "z_verbose" in zlib
+  - By [@fetzerch](https://github.com/fetzerch) in [#256](https://github.com/libretro/libretro-chailove/pull/256)
+
 ## 0.21.0 - 2018-05-02
 ### Chores
 - Switched to [libretro-deps](https://github.com/libretro/libretro-deps) for freetype and zlib